Clean up of documentation for NSDictionary implementation of factory
